rename: 슬라이더 파일 옮기기
