+++ src/Pages/Main/Main.tsx
@@ -1,9 +1,9 @@
 import { useRecoilState, useSetRecoilState } from "recoil";
 import { MainHeader } from "./MainHeader";
-import { StoreListSlider } from "./StoreListSlider";
-import { StoreStorySlider } from "./StoreStorySlider";
+import { StoreListSlider } from "./Slider/StoreListSlider";
+import { StoreStorySlider } from "./Slider/StoreStorySlider";
 import styled from "styled-components";
-import { StoreEditorSlider } from "./StoreEditorSlider";
+import { StoreEditorSlider } from "./Slider/StoreEditorSlider";
 import axiosInstance from "../../Api/axios";
 import {
   EditorDataState,
